+++ labs/RayTracer/src/Cylinder.cpp
@@ -15,7 +15,7 @@ void Cylinder::read(tinyxml2::XMLElement* element) {
     Figure::read(element);
 }
 
-void Cylinder::intersect_add_isect(vector<Isect> & intersecciones, Ray & ray, Vector & normal, Vector & point, float distance){
+void Cylinder::intersect_add_isect(vector<Isect> & intersecciones, Ray & ray, Vector & normal, Vector & point, double distance){
     Isect isect;
     isect.figure = this;
     isect.distance = distance;
@@ -29,10 +29,10 @@ void Cylinder::intersect_add_isect(vector<Isect> & intersecciones, Ray & ray, Ve
 }
 
 bool Cylinder::check_caps_intersect(vector<Isect> & intersecciones, Ray & ray, Vector & normal, Vector & point_ref){
-    float cosciente = normal * (point_ref - ray.origin);
-    float denominador = normal * (ray.direction);
+    double cosciente = normal * (point_ref - ray.origin);
+    double denominador = normal * (ray.direction);
     if (denominador != 0) {
-        float t = cosciente/denominador;
+        double t = cosciente/denominador;
         if (t>=Figure::FIGURE_EPS) {
             Vector point = ray.origin + ray.direction * t;
             if ((point - point_ref).Square() < radius*radius){
@@ -44,10 +44,10 @@ bool Cylinder::check_caps_intersect(vector<Isect> & intersecciones, Ray & ray, V
     return false;
 }
 
-bool Cylinder::check_body_intersect(vector<Isect> & intersecciones, Ray & ray, Vector & orientation, Vector & point, float tt){
+bool Cylinder::check_body_intersect(vector<Isect> & intersecciones, Ray & ray, Vector & orientation, Vector & point, double tt){
     Vector axis = (top - base);
     Vector r_point = (point - base).Projection(orientation);
-    float dd = r_point*axis;
+    double dd = r_point*axis;
     if ((dd > 0) && (axis.Square() > dd)){
         Vector normal = ((point - base) - r_point).UnitVector();
         intersect_add_isect(intersecciones, ray, normal, point, tt);
@@ -63,14 +63,14 @@ vector<Isect> Cylinder::intersect(Ray & ray) {
     Vector dp = ray.origin - base;
     Vector ray_p_orient = ray.direction.Rejection(orientation);
     Vector dp_p_orient = dp.Rejection(orientation);
-    float a = ray_p_orient.Square();
-    float b = 2*(ray_p_orient * dp_p_orient);
-    float c = dp_p_orient.Square() - (radius*radius);
-    float discrim = b*b - 4*a*c;
+    double a = ray_p_orient.Square();
+    double b = 2*(ray_p_orient * dp_p_orient);
+    double c = dp_p_orient.Square() - (radius*radius);
+    double discrim = b*b - 4*a*c;
     if (discrim >= 0){
-        float root = sqrt(discrim);
-        float t1 = (-b+root)/(2*a);
-        float t2 = (-b-root)/(2*a);
+        double root = sqrt(discrim);
+        double t1 = (-b+root)/(2*a);
+        double t2 = (-b-root)/(2*a);
 
         if (t1 > Figure::FIGURE_EPS){
             Vector point_1 = ray.origin + (ray.direction * t1);
