+++ labs/RayTracer/src/Ray.cpp
@@ -4,7 +4,7 @@
 Ray::Ray() {
 }
 
-Ray::Ray(Vector origin, Vector direction, float _tran) {
+Ray::Ray(Vector origin, Vector direction, double _tran) {
     this->origin = origin;
     this->direction = direction.UnitVector();
     this->tran = _tran;
@@ -22,9 +22,9 @@ Vector Ray::specularDirection(Vector incidentRay, Vector normal){
     return (incidentRay + normal * (incidentRay * normal * -2)).UnitVector();
 }
 
-bool Ray::transmisionDirection(float eta, Vector incidentRay, Vector normal, Vector & res){
-    float c1 = - (incidentRay * normal);
-    float cs2 = 1 - eta*eta * (1 - c1*c1);
+bool Ray::transmisionDirection(double eta, Vector incidentRay, Vector normal, Vector & res){
+    double c1 = - (incidentRay * normal);
+    double cs2 = 1 - eta*eta * (1 - c1*c1);
     if (cs2 < 0)
         return false;
     res = (incidentRay * eta + normal * (eta*c1-sqrt(cs2)));
