+++ labs/RayTracer/src/Ray.h
@@ -6,14 +6,14 @@
 class Ray {
     public:
         Ray();
-        Ray(Vector origin, Vector direction, float _tran=1);
+        Ray(Vector origin, Vector direction, double _tran=1);
         virtual ~Ray();
         Vector direction, origin;
-        float tran;
+        double tran;
         Vector rayPoint(const double dist);
         friend ostream& operator<<(ostream & , Ray & );
         static Vector specularDirection(Vector incidentRay, Vector normal);
-        static bool transmisionDirection(float eta, Vector incidentRay, Vector normal, Vector & res);
+        static bool transmisionDirection(double eta, Vector incidentRay, Vector normal, Vector & res);
     protected:
     private:
 };
